+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Next
+- Added jagged terrain generation
+
 ## 0.1.3_1
 *2025-03-02 17:52*
 - Include game version in save file